Move tests to test_table_functions.py
